requirements.txt
streamlit
openai
python-dotenv
pytesseract
pillow
PyMuPDF
pandas
matplotlib
azure-ai-formrecognizer
spacy
azure-storage-blob
